requirements.txt
Flask==3.0.3
Pillow==10.3.0
torch==2.3.0
ultralytics==8.2.28